% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{color_analysis}
\alias{color_analysis}
\title{Image plotter}
\usage{
color_analysis(x)
}
\arguments{
\item{X}{does a thing}
}
\value{
returns a dataframe of images
}
\description{
Function to extract image colors
}
\details{
the color class is a function that yields tertiary color regions, see https://en.wikipedia.org/wiki/Tertiary_color

a segmented version of colors could be available in the next zone
}
\examples{
colors() # This is unclear
}
