% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{convert_and_import}
\alias{convert_and_import}
\title{Convert and Import}
\usage{
convert_and_import(X)
}
\arguments{
\item{X}{Folder where images are stored}
}
\value{
returns a dataframe of images
}
\description{
Load all images into one
}
\examples{
load_images(here("Images/"))
}
