% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{imageplot_output}
\alias{imageplot_output}
\title{Image Plot Export}
\usage{
imageplot_output(Q, X, Y, A)
}
\arguments{
\item{X}{is the X var}

\item{Y}{is the Y var}

\item{A}{is the alpha}

\item{D}{is where the data is}
}
\description{
This function is designed to simplify passing arguments into a ggplot with geom_image to produce an image plot
}
\details{
function that allows you to pass alpha to a GG plot that also encodes other things
}
