% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{load_images}
\alias{load_images}
\title{Image Loader}
\usage{
load_images(y)
}
\arguments{
\item{y}{Folder where images are stored}
}
\value{
returns a dataframe of images
}
\description{
Load all images into one
}
\examples{
load_images(here("Images/"))
}
