% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{measure_images}
\alias{measure_images}
\title{Measure Image}
\usage{
measure_images(images)
}
\arguments{
\item{images}{Image to be read in}
}
\value{
Returns a dataframe called "measured_images" that is the meta data for the images and an OCR of image text
}
\description{
Measure image information and OCR (Optical character recognition)
}
\details{
This function returns a unified dataframe that takes your loaded images dataframe and returns a dataframe with image measurements and an OCR reading of the text from the image.
}
\examples{
measure_images(here("Images/image_1.png"))
}
