% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{thirds_images}
\alias{thirds_images}
\title{Rule of Thirds}
\usage{
thirds_images(images)
}
\arguments{
\item{images}{Folder where images are stored}
}
\value{
Returns a dataframe with scores for each third versus the image
}
\description{
Gets image symmetry
}
\details{
Detects intensity of the use of the thirds on a traditional photographic layout

Positive means there is more activity in the third versus the full image, negative means less
}
\examples{
symmetry(here("Images/"))
}
